requirements.txt
allure-pytest
requests~=2.32.4
pytest~=8.4.1
allure-python-commons~=2.14.3